Return TokenResponse from auth endpoint
I am just toying with this idea of returning a token as part of a json
object. The token would be one of two members, with the other being the
expiration timestamp. The point would just be to make it easier for the
client to know whether or not a given token is still valid, because it
can store the whole thing in localStorage and just check the timestamp
any time it pulls something out.

Of course, this makes my existing tests not work, because they expect to
receive a raw token from the server if they successfully authenticate.

Decisions, decisions.